core/requirements.txt
requests==2.17.3
pygal==2.3.1
pymongo==3.4.0
progressbar33==2.4